requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.9.0
django-crontab==0.7.1
django-extensions==2.2.8
django-filter==2.2.0
django-fontawesome==1.0
django-fontawesome-5==1.0.18
django-icons==1.1.2
django-meta==1.6.1
django-social-auth==0.7.28
djangorestframework==3.11.0
gunicorn==20.0.4
heroku==0.1.4
httplib2==0.17.1
idna==2.8
Markdown==3.2.1
oauth2==1.9.0.post1
oauthlib==3.1.0
plotly==4.6.0
portpicker==1.2.0
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==1.5
python-openid==2.2.5
python-social-auth==0.2.12
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
selenium==3.141.0
short-url==1.2.2
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==2.0
sqlparse==0.3.0
urllib3==1.24.3
webdriver-manager==2.3.0
whitenoise==5.0.1
wincertstore==0.2
django-crontab==0.7.1